Update README.md
Motivation

This project will give you experience in writing recursive functions that operate on recursively‐ defined data structures and mathematical abstractions.

Lists

A “list” is a sequence of zero or more numbers in no particular order. A list is well-formed if:

    It is the empty list, or

    It is an integer followed by a well‐formed list.

A list is an example of a linear‐recursive structure: it is “recursive” because the definition refers to itself. It is “linear” because there is only one such reference.

Here are some examples of well‐formed lists:

( 1 2 3 4 ) // a list of four elements ( 1 2 4 ) // a list of three elements

( ) // a list of zero element‐‐the empty list

The file recursive.h in the Project-2-Related-Files.zip defines the type list_t and the following operations on lists:

bool list_isEmpty(list_t list);

// EFFECTS: returns true if “list” is empty, false otherwise

list_t list_make();

// EFFECTS: returns an empty list

list_t list_make(int elt, list_t list);

    EFFECTS: given “list”, makes a new list consisting of

    the new element followed by the elements of the

    original list